Vinyl data fetcher working
