+++ package.json
@@ -4,12 +4,13 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
+    "dev": "npx index.ts",
     "build": "tsc && vite build",
     "preview": "vite preview"
   },
   "devDependencies": {
     "nodemon": "^3.1.7",
+    "tsx": "^4.19.1",
     "typescript": "^5.5.3",
     "vite": "^5.4.8"
   },
