--- src/index.ts
@@ -1,10 +1,0 @@
-const express = require("express");
-
-const app = express();
-const port = process.env.PORT || 3000;
-
-app.use(express.json());
-
-app.listen(port, () => {
-  console.log(`Server is running on port ${port}`);
-});
